LayoutTests/imported/w3c/web-platform-tests/content-security-policy/gen/top.http-rp/script-src-self/worklet-audio.https-expected.txt

CONSOLE MESSAGE: Refused to load https://www1.web-platform.test:9443/common/security-features/subresource/worker.py?redirection=keep-origin&action=purge&key=d686c70c-92f7-47a6-a697-5400a3a7a636&path=%2Fmixed-content because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://www1.web-platform.test:9443/common/security-features/subresource/worker.py?redirection=no-redirect&action=purge&key=59a66e1a-2f33-44ef-a59c-67f01a0a6d20&path=%2Fmixed-content because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://www1.web-platform.test:9443/common/security-features/subresource/worker.py?redirection=swap-origin&action=purge&key=a2c0fa58-d6b4-4cd6-8bb0-2c07aaefba55&path=%2Fmixed-content because it does not appear in the script-src directive of the Content Security Policy.

PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and keep-origin redirection from https context.
PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and keep-origin redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and no-redirect redirection from https context.
PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and no-redirect redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and keep-origin redirection from https context.
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and keep-origin redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and no-redirect redirection from https context.
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and no-redirect redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and swap-origin redirection from https context.
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and swap-origin redirection from https context.: securitypolicyviolation
FAIL Content Security Policy: Expects blocked for worklet-audio to same-https origin and swap-origin redirection from https context. assert_unreached: main promise resolved unexpectedly Reached unreachable code
FAIL Content Security Policy: Expects blocked for worklet-audio to same-https origin and swap-origin redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
